// resources/js/components/tmp/enums/todo.ts
export enum TODO_PANEL_STATE {
        SHOW = 'show',
        HIDE = 'hide',
}

export enum TODO_PANEL_MODE {
        UPDATE = 'update',
        CREATE = 'create',
}

export enum TODO_TASK_CATEGORY {
        ALL = 'all',
        FAVORITE = 'favorite',
        DONE = 'done',
        DELETE = 'delete',
}